started work on sharing topo infomation
